# openapi.yaml
openapi: 3.0.3
info:
  title: Users API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /api/v1/users:
    get:
      summary: List users
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
    post:
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"

  /api/v1/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      summary: Get user by id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: Not found
    put:
      summary: Update user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: Not found
    delete:
      summary: Delete user
      responses:
        "204":
          description: No content
        "404":
          description: Not found

components:
  schemas:
    UserCreateRequest:
      type: object
      required: [nombres, apellidos, rut, dv, fechaNacimiento, correoElectronico, contrasena]
      properties:
        nombres: { type: string }
        apellidos: { type: string }
        rut: { type: integer, format: int64 }
        dv: { type: string, minLength: 1, maxLength: 1 }
        fechaNacimiento: { type: string, format: date }
        correoElectronico: { type: string, format: email }
        contrasena: { type: string, minLength: 6 }
    UserUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/UserCreateRequest"
    UserResponse:
      type: object
      required: [id, nombres, apellidos, rut, dv, fechaNacimiento, correoElectronico]
      properties:
        id: { type: integer, format: int64 }
        nombres: { type: string }
        apellidos: { type: string }
        rut: { type: integer, format: int64 }
        dv: { type: string }
        fechaNacimiento: { type: string, format: date }
        correoElectronico: { type: string, format: email }